Correction of Actions and Reducers for feed and Reducer
